+++ CHANGES.md
@@ -135,6 +135,8 @@ Vernacular commands
 - The algorithm computing implicit arguments now behaves uniformly for primitive
   projection and application nodes (bug #9508).
 
+- `Hypotheses` and `Variables` can now take implicit binders inside sections.
+
 Tools
 
 - The `-native-compiler` flag of `coqc` and `coqtop` now takes an argument which can have three values: